+++ tests/Feature/Http/Controllers/NewsItemTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Feature;
+namespace Tests\Feature\Http\Controllers;
 
 use App\Models\NewsItem;
 use Illuminate\Support\Facades\Date;
