+++ tests/Traits/TestsFlashMessages.php
@@ -2,22 +2,21 @@
 
 declare(strict_types=1);
 
-namespace Tests\Feature;
+namespace Tests\Traits;
 
 use Spatie\Flash\Flash;
 use Spatie\Flash\Message;
-use Tests\TestCase;
 
-abstract class FeatureTestCase extends TestCase
+trait TestsFlashMessages
 {
     /**
      * Checks if the Spatie Flash message contains the given $message.
      */
     protected function assertFlashMessageContains(string $message): void
     {
+        /** @var Message $flashed */
         $flashed = app(Flash::class)->getMessage();
 
-        assert($flashed instanceof Message);
         $this->assertInstanceOf(Message::class, $flashed);
         $this->assertStringContainsString($message, $flashed->message);
     }
